refactor: move `CInstantSendDb` definition to `instantsend/db.{cpp,h}`
Also, since CInstantSendDb is used only by CInstantSendManager, we can
move it to the instantsend namespace easier.

Review with `git log -p -n1 --color-moved=dimmed_zebra`.
